api/requirements.txt
# FastAPI Backend Requirements

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1

# CORS
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Authentication (for future)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
